[Chore] 설명 주석 추가 / 함수,변수명 더 명확하게 수정

+++ src/components/Dashboard.jsx
@@ -10,28 +10,35 @@ import { ALERT_TYPE } from "../constant/constant";
 const { WARNING, SUCCESS } = ALERT_TYPE;
 
 export default function Dashboard() {
-  const pokemonList = useSelector((state) => state.pokemon);
   const dispatch = useDispatch();
-  const deleteAlert = createAlert({
+  const pokemonList = useSelector((state) => state.pokemon); // store에 저장된 포켓볼에 저장된 포켓몬들
+
+  // 삭제 재확인 팝업
+  const confirmDelete = createAlert({
     type: WARNING,
     content: "포켓몬을 삭제하시겠습니까?",
   });
 
-  const deleteSuccessAlert = createAlert({
+  // 삭제 성공 팝업
+  const successAlert = createAlert({
     type: SUCCESS,
     content: "포켓몬이 삭제되었습니다.",
   });
 
-  const initPokeBall = new Array(MAXIMUM_POKEMON - pokemonList.length).fill(
+  // 포켓몬이 들어있지 않은 포켓볼의 개수
+  const emptyPokeBalls = new Array(MAXIMUM_POKEMON - pokemonList.length).fill(
     null
   );
-  const myPokemonList = pokemonList.concat(initPokeBall);
 
+  // store에 저장된 포켓몬 + 포켓몬이 들어있지 않은 포켓볼 = 나만의 포켓몬 리스트
+  const myPokemonList = pokemonList.concat(emptyPokeBalls);
+
+  // 나만의 포켓몬 리스트에서 포켓몬을 삭제하는 이벤트
   const handleDelete = (targetPokemon) => {
     const { id } = targetPokemon;
-    deleteAlert().then((res) => {
+    confirmDelete().then((res) => {
       if (res.isConfirmed) {
-        deleteSuccessAlert();
+        successAlert();
         dispatch(deletePokemon(id));
       }
     });
